Reorganize resource edit and show components
