--- src/components/resource-edit/index.js
@@ -1,1 +1,0 @@
-export { default } from './resource-edit';
