--- src/components/resource-edit/resource-edit.css
@@ -1,12 +1,0 @@
-@import '@folio/stripes-components/lib/variables';
-
-.resource-edit-action-buttons {
-  align-items: flex-start;
-  border-top: 1px solid var(--minor-divider-color);
-  display: flex;
-  padding-top: 1em;
-
-  & button {
-    margin-right: 0.25em;
-  }
-}
