+++ src/components/resource/_fields/coverage-statement/coverage-statement-fields.js
