+++ src/components/resource/_fields/coverage-statement/index.js
