+++ src/components/resource/_fields/name/resource-name-field.css
