+++ src/components/resource/_forms/coverage-statement/index.js
