+++ src/components/resource/edit-managed/managed-resource-edit.css
